fix(surveillance): resolve alerts by their actual bigint id
BaseRepository<TEntity>.FindByIdAsync takes an int, but Alert.Id is a
long/bigint PK. Every caller that narrowed a long alertId down to int
before calling it threw ArgumentException at runtime (EF's DbSet.Find
validates the key CLR type against the model) — this affected the new
Surveillance ACL methods added in Phase 1, plus the already-shipped
AlertQueryService.GetAlertByIdQuery handler and every AlertCommandService
state-machine transition (confirm/dismiss/escalate/resolve/link-report/
mark-reviewed/add-timeline-record).

Add a long-typed FindByIdAsync overload to IAlertRepository/AlertRepository
and switch every call site to it. Surfaced while smoke-testing Phase 4's
specialist-matching endpoint end-to-end against a real alert.

+++ ArcadiaDevs.Viora.Platform/Surveillance/Application/Internal/CommandServices/AlertCommandService.cs
@@ -92,7 +92,7 @@ public class AlertCommandService(
     {
         try
         {
-            var alert = await alertRepository.FindByIdAsync((int)command.AlertId, cancellationToken);
+            var alert = await alertRepository.FindByIdAsync(command.AlertId, cancellationToken);
             if (alert is null)
             {
                 return new Result<long, Error>.Failure(SurveillanceErrors.NotFound);
@@ -139,7 +139,7 @@ public class AlertCommandService(
     {
         try
         {
-            var alert = await alertRepository.FindByIdAsync((int)command.AlertId, cancellationToken);
+            var alert = await alertRepository.FindByIdAsync(command.AlertId, cancellationToken);
             if (alert is null)
             {
                 return new Result<Unit, Error>.Failure(SurveillanceErrors.NotFound);
@@ -175,7 +175,7 @@ public class AlertCommandService(
     {
         try
         {
-            var alert = await alertRepository.FindByIdAsync((int)alertId, cancellationToken);
+            var alert = await alertRepository.FindByIdAsync(alertId, cancellationToken);
             if (alert is null)
             {
                 return new Result<Unit, Error>.Failure(SurveillanceErrors.NotFound);
